requirements.txt
absl-py==2.2.2
annotated-types==0.7.0
asgiref==3.8.1
attrs==25.3.0
cachetools==5.5.2
certifi
cffi==1.17.1
charset-normalizer==3.4.1
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
Django==5.1.6
dotenv==0.9.9
flatbuffers==25.2.10
fonttools==4.57.0
google-ai-generativelanguage==0.6.15
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
google-generativeai
googleapis-common-protos
grpcio
grpcio-status
httplib2==0.22.0
idna
jax==0.5.3
jaxlib==0.5.3
kiwisolver==1.4.5
matplotlib==3.8.4
mediapipe==0.10.21
ml_dtypes==0.5.1
numpy==1.26.4
opencv-contrib-python==4.9.0.80
opencv-python==4.9.0.80
opt-einsum==3.3.0
packaging==24.0
pillow==10.2.0
proto-plus
protobuf>=4.23.4,<5.0.0
pyasn1
pyasn1-modules
pycparser==2.22
pydantic==2.10.6
pydantic_core
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.31.0
rsa==4.9
scipy==1.12.0
sentencepiece==0.2.0
six==1.16.0
sounddevice==0.5.1
sqlparse==0.5.0
tqdm==4.66.2
typing_extensions  # Unpinned to allow >=4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1